Switching from Build from source to AppImage
At least in my test, this method of symbolic link doesn't work with the previous method where we built Musescore. So because of that, I am changing to download the AppImage. I believe that MuseSounds is more useful for most of the community than vst3 plugins (the patch will not be valid).  But this is my opinion!!!

What do you think?